refactor: Move device query functions from `speakers.py` -> `devices.py
Specifically default_speaker() and available_speakers()
